data type updates ||

+++ src/main/java/com/dbproject/chicagoincidents/domain/Quantitative.java
@@ -14,7 +14,7 @@ public class Quantitative {
     //enum: hole pots filled, carts delivered, premises baited, premises with garbage, premises with rats
     String quantity_type;
     @Nullable
-    Long quantity;
+    double quantity;
 
     @ManyToOne(fetch = FetchType.EAGER)
     @JoinColumn(name = "id", nullable = false)
@@ -23,7 +23,7 @@ public class Quantitative {
 
     public Quantitative() {}
 
-    public Quantitative(String quantity_type, Long quantity) {
+    public Quantitative(String quantity_type, double quantity) {
         this.quantity_type = quantity_type;
         this.quantity = quantity;
     }
@@ -44,11 +44,11 @@ public class Quantitative {
         this.quantity_type = quantity_type;
     }
 
-    public Long getQuantity() {
+    public double getQuantity() {
         return quantity;
     }
 
-    public void setQuantity(Long quantity) {
+    public void setQuantity(double quantity) {
         this.quantity = quantity;
     }
 }
